requirements.txt
fastapi==0.95.2
pydantic==1.10.8
pymongo==4.3.3
python-dotenv==1.0.0
uvicorn==0.22.0
watchgod==0.8.2
